Implement first pass at core game logic
This commit steps away from the webserver stuff and just focuses on the
core game logic. I've detailed in NOTES my thoughts on how the game
plays out. In particular, my first instinct was to reach to OOP to
create objects of cards, people, teams, games, etc. But I wanted to push
myself to try to create a super-minimal number based game (after all,
math is fun!).

One can deduce the value and suit of a card just from the value 0..23,
which opens nice doors into how to determine the winner of a trick (just
use the card value!).

I _believe_ the game state is simple enough to be held in a primitive
structure with "globals", and without the need for objects. I reserve
the right to double back and rewrite this in Moo though :)

Set up some basic testing to ensure the math is all correct.