Dealer: Add initial state tracking and player comms
This modifies the webserver to detect websocket clients on the '/play'
endpoint, and registers each with the Dealer, a new module that stores
the global state of the server.

It also adds the basic scaffolding for the client-server interaction
with the handle_msg dispatch sub.

As a first pass, we add the join_game action, where a client can join a
game, and this gets broadcasted to all others in the game.

~ckarle to test :)