Dealer: Add 'spectator' class of users for lobby-period
As discussed with ~ckarle earlier, to correctly implement the
seat-choice (team-choice) of a player, we really need to have two
classes of users in the game -- those seated and those not. I've deamed
them "players" and "spectators".

I think the correct order of items here will be to allow users to join
(to see the game state) if the game is NOT in progress. If the game is
in progress, we should prevent this (can be guarded by number of users,
or by a flag in the game state).

The flow will be:
                                    choose_seat
                                   .------------> player ------> start
                  join_game       /                | stand_up
    landing page ------------> spectator <---------'

The other routines will be implemented in future commits.