Dealer: Add sit/stand routines, with basic html test
This adds two new routines for the pre-game state management: take_seat
and stand_up. These two will be crucial for players to form teams,
moving from spectator status to a full player.

Some basic testing found it to be robust enough to push.

The other notable big change to state management is that the Empty seats
will all be 'undef' to simplify the standing up logic.