doc: Add inactivity-timeout note to README
We discovered this today -- Mojolicious closes the ws after 15 seconds
if it hasn't had any activity. This is obviously non-ideal for a game
with any thinking (or chatting haha).

While I look into how to put this in the server itself, I wanted to
document the workaround (via envvar) in the README