Euchre::Dealer: [refactor] store game obj in player hash
There's almost no time where we need the game_id in the player object.
It is much more likely (a.k.a literally every handler) that the first
thing we do is go get the game object out of %GAMES.

Storing the object reference in the player is just much cleaner.