Euchre::Dealer: [refactor] dispatch with player object
The handle_msg dispatcher used to send each handler the $cid of the
client (a.k.a their key in %PLAYERS). Just like the previous refactor,
we actually just want the player object 9 times out of 10. By adding the
id to the player object, we can clean up most references.