Euchre::Dealer: [refactor] Store player objects in game
This is another one of those revelations. I'm not sure _why_ I thought
it was a great idea to cross-link everything with the keys in the other
hash. AFAIK, it doesn't actually save any space (storing the object
reference is just like storing a pointer, which is probably cheaper than
the string!).

Maybe I first thought that broadcasting gamestate would get expensive if
you have the whole player objects in there. This isn't the case though,
because we map it to the player name before pushing it anyways!

Anyways, I tested this a bit and it doesn't seem to break anything. It
certainly makes things a bit cleaner.

Hey, maybe I'll break down and make these things actual objects instead
of hashes... maybe