Euchre::Dealer: Add trump_nominee to game state
After discussion with ~ckarle, we decided that it would be easier to
handle the "voting round" if the trump_nominee was a persistent part of
the game state rather than a single time value on the deal.

This will make it easier for clients to know what can and cannot be
nominated trump.