Euchre::Dealer: Add first pass at vote_trump action
This will need some tweaking in the near future. Particularly we need to
handle the 'loner' state. However, committing as is now as it (1) test
compiles and (2) isn't used anywhere.

There's an interesting mix appearing of server side validation vs client
side validation. I think my general leanings are towards the following
guidelines:

    * Communication Order -- handled by server (i.e. who can play a card
      when)
    * Communication Content -- handled by client (i.e. can you
      _actually_ play that card)

Obviously there is some content validation (valid messages) on the
server side, but otherwise, we don't manage hands or whether it's ok to
vote for what trump when. Makes things easier (and these are easy for
the client to handle themselves).