Euchre::Dealer: Make dealer track player hands
This is a large conceptual update in that we used to think we wouldn't
care about player hands.

The main motivation for this is to make the server robust to bad
connections. If you close a tab, refresh the page, etc, you'll need to
rejoin. The ONLY way to make this work is to keep track of the hand.

There are a few side benefits to this too:

    1) Simplifies writing clients (they don't need their own
       hand-management)
    2) Verifies clients don't cheat ;)

The actual overhead to make this happen is pretty minimal, so this is a
win all around.