Euchre::Dealer: Handle 'loner' value on order msg
The order msg (previously 'vote_trump') is used to vote for the trump
suit. It now passes the additional 'loner' boolean, which is used to
denote going alone!

To handle this properly / with the least confusion, I had to migrate the
table to 4 undefs and have people play into their "slot". This made
computing the winner idx much easier (no need to rotate the outcome of
trick_winner and accomodate whether someone sat out).

This has the added bonus of making it easier for clients to render each
slot on the table in the proper order.