Euchre::Game: Fix bug with trick_winner and static table
The previous commit made it so that p0 always played in c0 in the table,
which simplified the rotation logic and picking who goes next based on
who won. What I had FORGOTTEN was that trick_winner depends on the first
card in the list being the suit that was led!

The solution here is to start tracking the suit led. This is needed for
trick_winner to function properly; however, it also has the benefit of
making it easier to validate that players are following suit (both for
client and server).