lib/Euchre: [refactor] Change API to use char-style card/suits
Previously, the Euchre::Game subs all took int-style cards (i.e. suit
0-3 and card 0-23).

This was done before any work on the server, and it was fun to come up
with a minimal mathematical way to play the game.

In reality, this caused a lot of pain translating the cards between the
client websockets and server. Rather than put the onus on the server to
translate in communication with clients, the Game functions now need to
do the translation if they need the numeric value, and the Dealer never
knows what value the cards have (integer wise).

The Dealer deals with game state that now has human readable card and
trump values, and the Game logic handles the values if needed.

Tests have been updated accordingly. RIP Euchre::Card, you were too
complex.