Euchre::Dealer: Validate suit of card played
This commit adds validation logic to the suit of the card that is being
played. The Dealer (all knowing) inspects the hand of the player and
ensures that if they are NOT following suit, they CANNOT.

This is, of course, made more complex by the fact that the left bower is
the trump suit, but I think it came out clean enough logic wise.

The reason to validate server-side is that I don't plan to implement an
"undo" action any time soon, and so we want to prevent any user error
(or client error) from wrecking the game. It also means the clients can
be written quicker (~ckarle).