Euchre::Game: [refactor] Teach card_value about trump/led
This commit splits up card_value into card_value and raw_card_value. The
latter does what the original intended (i.e. give a number 0-23) and the
former now gives a number 0-98 based on trump/led.

This refactor simplifies trick_winner a lot, but that's not the main
motivation. The real reasoning here is the Dealer will, in a future
commit, take on the job of sorting hands. To sort easily, it needs to
know what's valuable. Why not use the same sub Game does!