Euchre::Dealer: Validate suit being ordered
We know the trump nominee; we know the suit you're requesting; we know
how many times it's passed. Might as well yell at you if you're
misbehaving! (The Dealer is getting awfully good at that these days...
so much for the laid back "if you say so" Dealer :) )

Also adds a <select> element to debug.html to test this nonsense.