Euchre::Dealer: Add dealer_swap phase after ordering trump
This is the final phase in the strange and twisted dance that is
ordering trump. If you're in the first round of voting, you MUST vote
for the kitty card and that kitty card will go to the dealer.

To ensure that the dealer gets it, we add a new "dealer_swap" phase,
blocking all other players from doing anything (I knew the handle_msg
generic validation would come in handy!). We then wait for the dealer to
give us a card from their hand to exchange, and give them the
trump_nominee in return.

Smooth as butter. Added click handlers to debug.html to test it.