Euchre::Dealer: Add force join_game option to take over hand
This is crucial to support dropped players. While it is definitely prone
to "euchre.live bombing" (in the parlance of our times), where someone
could guess a player+game combo and take over the hand, we'll trust our
beta testers (read: family) for now to be good citizens.

The mechanism works by adding a force flag to join_game. If present, and
if the username is already in the game, we hot-swap the player's WS
object and ID. Name/hand/game all stay the same. Works for both players
and spectators :)