Euchre::Dealer: Add prev_table to game state for clients
When the last card of the trick is played, the game keeps chugging,
sending each player an empty table, meaning NO ONE knows what that
person actually played! (are they cheating? was that an Ace? Who counts
trump anyways?).

The easiest solution here, short of adding an async pause to the server,
is to include the previous table in the game state. Clients can choose
to always display it (cough: my lazy debug page), or they can choose to
detect that the table is empty and temporarily display the previous
table for a few seconds before allowing users to vote/play cards.

In the process, I removed the clearing of table in start_new_round; it
was redundant (set in play_card or join_game).