Euchre::Dealer: Use timer post-trick over prev_table
The previous, hacky, logic for having the clients show the table after
the trick was done was to include the table as prev_table forever.

This commit scraps that logic in favor of a 'pause' phase with
mojolicious' Mojo::IOLoop Timer, which is 100% cleaner and what we
wanted in the first place!

Who knew it pays to read the docs?