ci: Add build script and release recipe
After a few times, this got super painful to do by hand.

The current game plan is to do all the webpack building on the local dev
machine, as the VM has only 1G RAM and runs out of memory compiling the
scss (yikes).

Luckily, the whole idea of webpack is that it doesn't need to rebuild if
the JS doesn't change, so for our production env, we only need a subset
of the repository.