Euchre::Dealer: Add resource cleanup on players leaving
Previously there was absolutely 0 cleanup of the %GAMES variable. This
is, by definition, a nice memory leak :) and the poor 1GB RAM of VM
space would slowly be chipped away at as the uptime grew.

Thankfully, Mojo::IOLoop made it pretty easy to add a check after a
player has left to ensure that the game is cleaned up properly. It goes
like this:

    On player leaving:
        If they're in a game:
            Make note to delete game no players are active after 30 min

Players "leave" when they disconnect for any reason. Tab closing,
dropped connection, refreshed tab, whatever. In these cases, we delete
the reference to the player in our global %PLAYERS state, but we do NOT
delete the object in the $game->{players}.

This is intentional, as we want to allow that player to rejoin the game,
in case of network disruption / accidental tab closage. They can rejoin,
and we'll swap them in and mark them as active again.

So as long as no player is active in the game (has the tab open and
pinging our server), we'll delete the game after 30 mins of inactivity.

Tracking "activity" is as easy as assuming they're active when they
register and marking inactive when they leave. When we swap the player
in, we take most of the player's previous values, so we mark the old
player active. No other actions change activity (that's a mouthful..
it's late).