Euchre::Dealer: Enhance stats with times and names
This commit introduces the notion of "joined" and "started" times for
players and games respectively. This is crucial for the stats to track
so that I can know when someone has just gone AFK for days.

Additionally, it shows the screen names of users and the games they're
in. This isn't really there to "track" anyone; it's just a small
community of players (beta testers) right now, and I'd like to know who
to contact if I need to restart the server.

Longer term, it's also just easier to remember usernames that haven't
left for days instead of UUIDs...