Euchre::Dealer: Add leave_game action for endgame
In theory, this is doable at any time, but we'd really only like people
to do this in lobby/end for now (a future commit will hopefully change
it).

Anyways, the critical thing is that we take the player out of the $game
so that the gamestate is no longer sent to them. It'll be up to the
client to at this point redraw a start screen or something (does this
make debug.html no longer a valid client? it certainly is a hack...)