Euchre::Dealer: Add basic password-protection to games
The first person to join the game has the option to join with a
password, which each additional player will have to specify in order to
join.

This is needed to prevent "euchre-live bombing" (a-la Zoom-bombing)
where a player could see that Alex was playing at table Foo and could
force join to take over the hand.

Having a trusted environment for each table also means that we can push
everyone into a spectator role in end game and no one should be worried
about seats stolen.