Euchre::Dealer: Convert all players to spectators at end
When the game is over, we move all players to the spectator role. If
they want to play another game, they can pick their seats again. If they
want to exit, that's cool too. The important thing is that we don't want
any grandpa's walking away from a screen with an open tab preventing
someone who WAS spectating from taking that seat :)