refactor: Make Game and Player hashes objects
(spoiler alert!) I'm planning on adding a third major type: Tables, but
before I added another hash in the fray, I figured I should PROBABLY
move them all to bonefide objects.

The perks here are:

    1. Can't accidentally create new properties
    2. Can move a bunch of the actions that act on hashes to methods

Already I've ironed out a few small bugs from the leniency of hashes.

I don't particularly enjoy object oriented programming, but it just
feels right for this problem.

I chose Class::Tiny as my object framework because (1) it's only a
hundred or so lines of pure perl with no dependencies and (2) I really
don't need the crazy inheritence, encapsulation, and other things that
come with Moose or even Moo.

This commit mostly is the syntax changes to move the hash keys to proper
object attributes (getters and setters). Not many of the subroutines
have become methods yet (only send_error as an example). More on the
latter to come.