Euchre::Host: Demote the Dealer, long live the Host
So all this work on OOP cleanup has been to support three things:

    1. No circular object dependencies
    2. Clearer separation of roles
    3. Stricter accessors to prevent bugs

So why split the Dealer's responsibilities between a new Host class and
the existing Dealer (now with many Dealers)? Well the OOP hierarchy
looks like this:

    gloat.pl (server)  ----.
                           |
                           v
                         Host (table manager)
                       /   |   \
                      v    v    v
                  Dealer Dealer Dealer

Where each Dealer represents a conceptual "Table" (Players and Game).

So the Host is responsible for getting you to a "Table" (join / create /
list / leave tables), which is done by registering you with the Dealer
associated with that Table.

The Table doesn't really exist as an object, more of a concept. The name
Dealer was kept because the core reason to have a Table is to allow for
communication within that group without crazy cross referencing. It felt
weird to have a Table talking to you :) (it is pretty late at night...)

So the main benefit here is that the Game no longer needs to know about
Players (at all). It doesn't even need to know what seat is calling it
-- it already knows who's turn it is and the Dealer ensures Players take
turns fairly.

Likewise, Players are now free to sit/stand willy nilly, as the Game
holds the hands. This simplifies the "dropped connection" workflow, as
we don't need to the whole "force join" concept. If a connection drops,
remove the player and they (or any other spectator) can sit back down
and resume.