lib/Euchre: Add BUILD subs to classes to de-lazify attrs
No one likes a lazy "start_time" attribute ;)

This is a clever hack to de-lazify them. The BUILD sub is supposed to be
used for validation at object creation time, but we can just as easily
use it for the instantiation of some key fields which our clients might
be expecting (i.e. score needs to be present before the first round
finishes).

~ckarle -- how do you like them oneline for loops :P