debug.html: Remove Force Join button
This is no longer a supported action. With the separation of Players and
hands, we can have players drop in and out willy nilly and not lose game
state. As such, there's no reason to support forcefully joining a game
and booting a current player.

If someone loses a connection, we kick them immediately and they can
still then join under the same name.