Euchre::Errors: Add safeguards to ensure keys present in msg
We had a bug during our playthrough tonight that would have been more
obvious if the server had failed loudly at the required key not being
present in the message.

This commit adds those safeguards, with a check for required fields
prior to handling any other requirements or doing any other work.

This should ideally be transparent to end users, and only be used so
thta we don't enter a "cascading bad" state if a client sends a poor
message.