Euchre::Host: Delete finished games on last leave_game
When a Player exits (either via leave_game themselves or by closing a
tab and killing the connection), we check if the Table they were at has
a game that is finished. If they're the last Player at the Table to
leave, AND if the game was over (i.e. no progress), we kill the table.