make: Add proper dependency logic to build/release
As mentioned in the top comment, building webpack take both over a
minute on my development laptop and gets my fans spinning real good.

So for cutting a release where no public content has changed, it's
reallllly painful to have to sit through that! (ok, I exaggerate, but my
hobby time is precious!)

Thankfully, this is exactly what make was built for -- only building
that which has changed. Throw in a couple `find`'s and some dependency
magic with timestamp files and boom, we don't have to rebuild webpack
anymore!