Euchre::Host: Add Players table to /stats
There have been some leaked players recently, so it's helpful to see
both players in the games and players not in games (that may have been
leaked).

In particular, it appears closing a laptop lid doesn't properly close
the websocket... or at least send the "finish" message properly. I'll
have to investigate further.