Euchre::Host: Add prune_players to catch stale connections
If a client suspends their device (laptop lid, phone screen), it appears
the connection isn't properly terminated.

I was under the impression we should have detected this when we stopped
sending them ping/pong actions... but I guess not.

So this commit adds a second resource cleanup sub -- prune_players.
Thankfully, it's easy to check if the ws is open, so we just need to go
through and delete all PLAYERS with stale ws connections. This is safe
because if the ws thinks it's closed, it certainly aint gonna reopen
(not with our clients at least).

Also a minor syntax update to the file to use the id accessors over the
{} hash style access. Hard to break that habit.