ChatPanel
Fix a bug with chat scrolling.  It was working beautifuly but then
I had the audacity to put a border on the outer div and all hell
broke loose, with the panel freezing and content getting added to the
bottom, breaking the bounds and lengthening the whole page.

Adding a border broke the scrolling.  F*@king CSS.