preprod: Add a new preprod.html for our staging site
I'm configuring our nginx reverse proxy to map stage.euchre.live to our
preprod version of the site, so that we can move fast and break things
while allowing relatives to play all the while.

This commit sends over a modified frontend with a warning banner should
the user be using the preprod version (so that we can easily
distinguish).