make: Add version.txt artifact to build area
With the rollout of preprod and prod, we need an easy way to know what
version the server is running. I thought about baking in a version
number in the server itself, but it's much easier and more informative
to just save off the git log from the HEAD at the time the release is
built.

Can be retrieved with a simple GET /version.txt