make: Update release dest, check in restart script
This is the infamous "safe deployment" script I've been using for a
while, tuned up for the preprod vs prod deployment. It basically boils
down to using pgrep/pkill to kill the old server (if running, and if no
players are on the server), then starting the server at the right port.

So the general workflow is to run "make release" to deploy to preprod,
and to cut a prod release, you need to ssh into the server, replace
"prod-el" with "preprod-el" and then call "restart.sh" manually.

A future commit will trim down some of these steps... but for now I
wanted to check in the WIP so that it wasn't lost (if I jump VPS
providers any time soon).