make: Build production webpack JS on `build` target
Previously, we only built the development version. It came to my
attention today that the development version is 8MB and the production
version is 400KB, so we should really be using production.. in
production (surprise?).

This commit updates the `make build` recipe to build the production
webpack bundle, which has been renamed to be just webpack.production.js
(to allow it to be referenced statically).

The HtmlWebpackPlugin previously in use was actually generating a
webpack.production.html that referenced webpack.SHA.js, but for now
(MVP), we'll stick to our static index.html. (This generation seems
necessary for Mojolicious::Webpack::Plugin though)