ci: Add deploy stage to deploy to preprod
This leverages the secrets feature of builds.sr.ht (for which I have
added a SSH key and put its pair on the server) so that I can cut a
release on a push!

This is safe to run, as it (1) only tries to restart preprod and (2)
will fail if the server has players.