ci: Temporarily reduce CI to just setup/test
The npm step for the build fails every so often with the following:

: extract:carbon-components: sill extract carbon-component
: makR! cb() never called!

So while I investigate, I'm disabling the build/deploy part and we'll go
back to manual releases (with just the tests running in CI).

This will also have the perk of way speeding up our build time.