make: Minimize reasons to rebuild with webpack
Updating public/*.html doesn't require rebuilding all the javascript we
ship (usually).

As such, I'm updating the reasons to run build.sh to be JUST the ASSETS
in assets/, everything else under public is a reason to rebuild the
build/ directory, but not a reason to rebuild build/asset.

Long live fast deployments.