prod: Fix production server daemon behavior
So I was under the impression that the 'daemon' command, well, created a
daemon. This is false.

To truly daemonize, one needs to call the server's daemonize command
(which took forever to figure out how to do, due to the class
hierarchy... but I figured it out!).

Anyways, with that solved, the only thing left was how to still capture
logs. The solution is to use a global logger (i.e. shared settings)
between all modules (tied to the $:: a.k.a. main namespace).

With both of these in place, the rc.d script finally looks normal. After
all, rc.d is about managing daemons, not processes that want to be
daemons :)