deps: Update node dependencies, Mojo::Webpack
The Mojo::Webpack upgrade was accidental (new machine, new release in that
time frame). With it came some changes to the generated webpack.config.js
that threw me for a loop for a good while.

Most importantly, a newer version of node-sass was needed. As such, this
patch updates all dependencies.